requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dask==2024.5.1
db-dtypes==1.2.0
fsspec==2024.5.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
locket==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
partd==1.4.2
polars==0.20.26
proto-plus==1.23.0
protobuf==4.25.3
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
toolz==0.12.1
tzdata==2024.1
urllib3==2.2.1
